requirements.txt
flask==2.0.1
google-genai
python-dotenv==0.19.0
flask-cors==3.0.10
flask-socketio==5.1.1
python-socketio==5.10.0
gunicorn==21.2.0
Werkzeug==2.0.1
bidict==0.22.1
click==8.1.7
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
simple-websocket==1.0.0
wsproto==1.2.0
python-magic==0.4.24
